Sprint2 completed

+++ src/lib/type_cast.hpp
@@ -38,7 +38,7 @@ const T& get(const AllTypeVariant& value) {
 // Template specialization for everything but integral types
 template <typename T>
 std::enable_if_t<!std::is_integral<T>::value, T> type_cast(const AllTypeVariant& value) {
-  if (value.which() == detail::index_of(types, hana::type_c<T>)) return get<T>(value);
+  if (static_cast<size_t>(value.which()) == detail::index_of(types, hana::type_c<T>)) return get<T>(value);
 
   return boost::lexical_cast<T>(value);
 }
@@ -46,7 +46,7 @@ std::enable_if_t<!std::is_integral<T>::value, T> type_cast(const AllTypeVariant&
 // Template specialization for integral types
 template <typename T>
 std::enable_if_t<std::is_integral<T>::value, T> type_cast(const AllTypeVariant& value) {
-  if (value.which() == detail::index_of(types, hana::type_c<T>)) return get<T>(value);
+  if (static_cast<size_t>(value.which()) == detail::index_of(types, hana::type_c<T>)) return get<T>(value);
 
   try {
     return boost::lexical_cast<T>(value);
